Restructure keymap to better match the approach adopted by the file browser
